v1/requirements.txt
pandas
matplotlib
faiss-cpu
sentence-transformers
fastapi
uvicorn